Gate E2E-only session metadata command on the smoke-test driver
