+++ src/vs/workbench/contrib/languageRuntime/browser/languageRuntime.contribution.ts
@@ -5,7 +5,7 @@
 
 import { Registry } from '../../../../platform/registry/common/platform.js';
 import { registerLanguageRuntimeActions } from './languageRuntimeActions.js';
-import { registerSessionMetadataCommand } from './sessionMetadataCommand.js';
+import { registerLanguageRuntimeActionsForSmokeTests } from './languageRuntimeActionsForSmokeTests.js';
 import { PositronRuntimeLanguagesContextKeyContribution } from './languageRuntimeContextKeys.js';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from '../../../common/contributions.js';
 import { LifecyclePhase } from '../../../services/lifecycle/common/lifecycle.js';
@@ -14,9 +14,8 @@ import './languageRuntimeActions.css';
 // Register commands
 registerLanguageRuntimeActions();
 
-// Register the internal command that returns session metadata for
-// programmatic callers (e.g. the e2e smoke-test driver).
-registerSessionMetadataCommand();
+// Register commands that are only available when the E2E smoke-test driver is enabled
+registerLanguageRuntimeActionsForSmokeTests();
 
 // Register the contribution that tracks language IDs with a registered
 // Positron runtime, so menus/actions can scope to "this editor's language
